+++ src/main/java/http/exceptions/InvalidSessionException.java
@@ -1,4 +1,4 @@
-package http.session;
+package http.exceptions;
 
 public class InvalidSessionException extends RuntimeException {
     public InvalidSessionException() {
